+++ i18n/config.json
@@ -1,0 +1,18 @@
+[
+  {
+    "code": "en",
+    "hrefLang": "en-US",
+    "name": "English",
+    "localName": "English",
+    "langDir": "ltr",
+    "dateFormat": "MM/DD/YYYY"
+  },
+  {
+    "code": "fr",
+    "hrefLang": "fr-CA",
+    "name": "French",
+    "localName": "French",
+    "langDir": "ltr",
+    "dateFormat": "DD.MM.YYYY"
+  }
+]
