+++ i18n/react-intl/en.json
@@ -1,0 +1,3 @@
+{
+  "helloWorld": "Hello World"
+}
