+++ i18n/react-intl/fr.json
@@ -1,0 +1,3 @@
+{
+  "helloWorld": "lolz les copains"
+}
