+++ package.json
@@ -19,11 +19,14 @@
     "gatsby-plugin-styled-components": "^3.3.10",
     "gatsby-source-contentful": "^2.3.29",
     "gatsby-source-filesystem": "^2.3.14",
+    "gatsby-theme-i18n": "^1.0.2",
+    "gatsby-theme-i18n-react-intl": "^1.0.2",
     "gatsby-transformer-sharp": "^2.5.7",
     "node-sass": "^4.14.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-helmet": "^6.1.0",
+    "react-intl": "^5.4.2",
     "styled-components": "^5.1.1"
   },
   "devDependencies": {
